+++ server/readResults.js
@@ -1,3 +1,0 @@
-let db = require('./router-dinesh');
-
-console.log(db.findAll("","",""));