+++ src/views/Results/Results.js
@@ -3,25 +3,49 @@ import React, {Component} from 'react';
 import SearchResultList from './app/components/SearchResultList';
 //import * as dbIntegration from './../../../services/dbIntegration';
 
-import accessserver from './accessserver/';
+import axios from 'axios';
 
 import {Card, CardBody, CardHeader, Col, Pagination, PaginationItem, PaginationLink, Row, Table} from 'reactstrap';
 
-class Tables extends Component {
+class Results extends Component {
 
   constructor(props) {
     console.info('Inside constructor');
     super(props);
-    this.state = {
+    /*this.state = {
       searchResults:{
         fromDate: '2018-10-22 00:00:00',
         toDate: '2018-10-24 00:00:00'
       }
+    };*/
+
+    this.state = {
+      posts: [],
+      loading: true,
+      error: null
     };
   }
 
+  findDineshResults(){
+    axios.get('https://jsonplaceholder.typicode.com/users')
+      .then(res => {
+        const posts = Object.entries(res.data).map(obj => obj);
+        this.setState({
+          posts: posts,
+          loading: false,
+          error: null
+        });
+      })
+      .catch(err => {
+        this.setState({
+          loading: false,
+          error: err
+        });
+      });
+  }
+
   componentDidMount() {
-    //this.findProducts();
+    this.findDineshResults();
     console.log('Inside Component Mount');
   }
 
@@ -51,7 +75,7 @@ class Tables extends Component {
           <Col>
             <Card>
               <CardHeader>
-                <i className="fa fa-align-justify"></i> Bordered Table
+                <i className="fa fa-align-justify"></i> Search Results
               </CardHeader>
               <CardBody>
                 <Table responsive bordered>
@@ -63,17 +87,8 @@ class Tables extends Component {
                     <th>Status</th>
                   </tr>
                   </thead>
+                  <SearchResultList results={this.state.posts} test="Dinesh"/>
                 </Table>
-                <Pagination>
-                  <PaginationItem><PaginationLink previous tag="button">Prev</PaginationLink></PaginationItem>
-                  <PaginationItem active>
-                    <PaginationLink tag="button">1</PaginationLink>
-                  </PaginationItem>
-                  <PaginationItem className="page-item"><PaginationLink tag="button">2</PaginationLink></PaginationItem>
-                  <PaginationItem><PaginationLink tag="button">3</PaginationLink></PaginationItem>
-                  <PaginationItem><PaginationLink tag="button">4</PaginationLink></PaginationItem>
-                  <PaginationItem><PaginationLink next tag="button">Next</PaginationLink></PaginationItem>
-                </Pagination>
               </CardBody>
             </Card>
           </Col>
@@ -83,4 +98,4 @@ class Tables extends Component {
   }
 }
 
-export default Tables;
+export default Results;