+++ src/views/Results/app/components/SearchResultList/SearchResultListItem/index.js
@@ -1,0 +1,11 @@
+/* eslint-disable */
+import React from 'react';
+
+class SearchResultListItem extends React.Component {
+
+    render() {
+        return ("<tr> <td>1</td><td>dinesh</td><td>one</td><td>two</td></tr>");
+    }
+};
+
+export default SearchResultListItem;