+++ src/views/Results/app/components/SearchResultList/index.js
@@ -3,13 +3,13 @@ import React from 'react';
 import SearchResultListItem from './SearchResultListItem';
 
 class SearchResultList extends React.Component {
-  render() {  
-    let listItems = this.props.results.map(result =>
-      <SearchResultListItem key={result.id} result={result}/>
+  render() {
+    let listItemsActual = this.props.results.forEach(([key, value]) =>
+      <SearchResultListItem key={key} result={value}/>
     );
     return(
       <tbody>
-        {listItems}
+      {listItemsActual}
       </tbody>
     );
   }